Addressing failures in Daily Runs (#81)
This PR addresses failures in the daily runs as well as adds missing failed block tests for DC.

Most of the tests in the daily runs were failing due to internal network access denied issues and browser bot issues.  These tests are being skipped temporarily until a solution is in place that will work.